# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/sale_order": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of sale orders or a specific sale order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_orders"
                ],
                "summary": "Get sale orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order ID",
                        "name": "sale_orderID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.SaleOrder"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new sale order with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_orders"
                ],
                "summary": "Create a new sale order",
                "parameters": [
                    {
                        "description": "Sale Order",
                        "name": "sale_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrder"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sale_order/frontend": {
            "post": {
                "description": "Create a new sale order with customer info and cart items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_orders"
                ],
                "summary": "Create a new sale order from frontend",
                "parameters": [
                    {
                        "description": "Sale Order Request",
                        "name": "sale_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrderRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Anonymous token (optional)",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrder"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sale_order/{sale_orderID}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an sale order with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_orders"
                ],
                "summary": "Update an sale order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order ID",
                        "name": "sale_orderID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sale Order",
                        "name": "sale_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrder"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an sale order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_orders"
                ],
                "summary": "Delete an sale order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order ID",
                        "name": "sale_orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale Order deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Sale Order not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CartItem": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "types.SaleOrder": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "types.SaleOrderRequest": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CartItem"
                    }
                },
                "customer_address": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "customer_phone": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}